requirements.txt
accelerate==1.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
einops==0.8.1
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.6.1
huggingface-hub==0.29.1
idna==3.10
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.4
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==2.1.5
matplotlib==3.9.4
mpmath==1.3.0
natsort==8.4.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.11.0.86
packaging==24.2
pillow==11.0.0
psutil==7.0.0
pycocotools==2.0.8
pyparsing==3.2.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scikit-image==0.24.0
scipy==1.13.1
six==1.17.0
sympy==1.13.1
tifffile==2024.8.30
tokenizers==0.21.0
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
urllib3==2.3.0
zipp==3.21.0
